Create vimsetup.sh
I created this because I felt too much time was spent on my end setting up shellcheck each time I installed vim for a dev environment - then modifying other settings. 

This will save an admin or programmer a lot of time and I encourage any and all to tweak it for subjective preference. As a sociologist professor once jokingly said, "what's good for the goose is not always good for the gander." Meaning? What I dig - well, it might not be your cup of tea.